(Staticman) Stephen Cleary: In that example, the `result.Add(data)` lines may run in parallel.
TAP methods return "hot" (in progress) tasks. So consider `GetBothAsync`. It calls `GetOneAsync`, which starts that method and returns a task. `GetOneAsync` is in progress at this point. Then it calls `GetAnotherAsync`, which starts that method and returns a task. `GetAnotherAsync` is in progress at this point. So both methods are running concurrently.

https://blog.stephencleary.com/2017/03/aspnetcore-synchronization-context.html#comment-c4170cd0-d8a1-11e9-ad4c-b516289148dc