(Staticman) Stephen Cleary: Yes, that does prevent deadlocks. It still blocks a thread, though, since it uses `Wait`/`Result`. Also - as a side note - exception handling is different when using `Wait`/`Result`; use `GetAwaiter().GetResult()` if you want to avoid `AggregateException` wrappers.
The reason this avoids the deadlock is that there's no context captured. `Task.Run` runs its delegate without a context, so any `await`s in `CallHttp` will not capture a context. But it does block a thread, so it's better to convert `Get` to use `async` and `await` if possible.

If that's not possible and you *need* to mix synchronous and asynchronous code, see my article on [Brownfield Async](https://docs.microsoft.com/en-us/archive/msdn-magazine/2015/july/async-programming-brownfield-async-development).

https://blog.stephencleary.com/2012/07/dont-block-on-async-code.html#comment-09126190-1b73-11ea-91f4-bd5a4438eddf