(Staticman) Stephen Cleary: ASP.NET Classic was certainly the biggest offender; due to limitations in the framework, blocking on asynchronous code was sometimes necessary for MVC child actions and MVC action filters. ASP.NET Core both removed the context *and* made the entire pipeline asynchronous, so there's nowhere where developers have to block on asynchronous code anymore.
In the broader .NET Core world, the only other common place where a context is found is in UI applications. And in that particular case, you almost never want to block the UI thread *anyway*. Also, in a UI app, if you *really do* need to block the UI thread, you can shove the work onto a background thread using `Task.Run` (you wouldn't want to do this in a server situation like ASP.NET because it would limit your scalability).

So, there aren't any scenarios remaining where `ConfigureAwait(false)` is necessary to avoid deadlocks. But on the other hand, that was not what it was actually intended for in the first place *anyway*. `ConfigureAwait(false)` should be used in general-purpose library code that doesn't need to resume on its context.

There's currently a movement to do away with `ConfigureAwait(false)` completely. For a specific application or private library, that may make sense. I'm not convinced that's the best approach for a general-purpose library; that would be a judgement call, depending on the library and its intended usage.

https://blog.stephencleary.com/2017/03/aspnetcore-synchronization-context.html#comment-251cc460-1b74-11ea-91f4-bd5a4438eddf