(Staticman) Stephen Cleary: This is due to the ASP.NET request context. When a request arrives at ASP.NET, it creates a request context (including `HttpContext.Current` as well as a few other things) and then calls `GetValue1`.
When your action does its `await`, it captures its current context (in this case, `SynchronizationContext.Current`, which is an instance of `AspNetSynchronizationContext`) and then returns to its caller. ASP.NET then allows this thread to return to the thread pool.

Later, when `ReadToEndAsync` completes, a thread is taken from the thread pool and the action is resumed *within that context* - the `AspNetSynchronizationContext` sets `HttpContext.Current` (and a few other things) before it resumes executing `GetValue1`.

So it can change threads, but it's the same request context.

https://blog.stephencleary.com/2012/02/async-and-await.html#comment-c9ebc6d0-1b74-11ea-91f4-bd5a4438eddf