(Staticman) Stephen Cleary: > Let's say I have a method that does both IO-bound and CPU-bound work.
That's the subject of [the next blog post in this series](https://blog.stephencleary.com/2013/11/taskrun-etiquette-examples-even-in.html).

> do you have any suggestions or best practice how to make this behavior clear to the caller?

I definitely recommend clear XML comments for methods like this. You might also want to consider a naming convention; there isn't one in common use but you can establish one for your own codebase.

As a side note, this is a very rare scenario. "CPU-bound work" in this context isn't just parsing HTML or anything simple like that; it refers to a substantial amount of work to the point where it would be noticeable and undesirable when running on a UI thread. [50ms is a good rule of thumb](https://blog.stephencleary.com/2013/04/ui-guidelines-for-async.html), though in my experience you could double that and still be good.

> I understand that, if an IO-bound async method is well implemented, it does NOT block ANY thread (or at least not one from the pool).

Correct; it doesn't block a thread for the duration of the operation.

> because the first example blocks the thread from Task.Run, whereas the second example does not?

Yes; the first example blocks a thread on the thread pool. That thread is blocked on the call to `Result`. The second example uses a thread pool thread to start the asynchronous operation, but does not block a thread pool thread until it completes.

https://blog.stephencleary.com/2013/11/taskrun-etiquette-examples-dont-use.html#comment-69b54640-1b76-11ea-91f4-bd5a4438eddf