(Staticman) Ohad Tsamir: Hi Stephen, In my company, we've been using CallContext within WCF as a replacement for OperationContext.Current, which could become null after async operations. I've recently stumbled upon a behaviour that alarms me: Sporadically (more often with high concurrency), a request calling `LogicalGetData` will gain access to data stored via a **different** request's call to `LogicalSetData`. Would you have any explanation to this? Could execution context be shared between different requests to a WCF service hosted in IIS?
If relevant, the calls to both methods are performed via a `MessageInspector`.

Additionally: The original data I stored was not immutable (because I only read your post now :) ), so I created a new test case where I only store the string `OperationContext.Current.IncomingMessageHeaders.Action`, which in WCF holds the operation name, and indeed I could still (sporadically) access another operation's action. But I'm wondering if this is of any relevance? If I understand your post correctly, the importance of immutability is to ensure that the data is in the correct state, but I'm assuming that even mutable objects shouldn't be able to cross the boundaries of a single execution context, correct?

[Here](https://stackoverflow.com/a/59787566/198797)'s an answer I recently posted on SO (feel free to correct / address me either here or there).

https://blog.stephencleary.com/2013/04/implicit-async-context-asynclocal.html#comment-ecccbe80-3941-11ea-a228-db01b2ef7800