(Staticman) nowave7: Hey Stephen,
First off, thanks for this post. It sure did clear up quite a few questions I had regarding this topic.

But, there are a few remarks I'd like to make and a few questions I'd like to ask.
1. When you say that there is no thread firing the I/O request, that's not entirely true, since you have at least that original UI thread that initiated the I/O operation in the first place? Of course, your argument a bit further down the line, that there is no new thread being created that will block while executing the request is a whole different statement. Am I understanding this correctly? I think it is important to be as precise as possible here, exactly to avoid any kind of confusion.
2. Now onto something that is still not clear to me.
What exactly happens to the UI thread (or the application itself) during the propagation of the I/O request all the way down to the device itself (or to be more precise till the point when the BCL/library returns an incomplete task). I'm assuming it remains unresponsive? Even for no longer than (and I'm guessing here) 10ms?
3. And finally, you say that the BCL has already registered a handle with the I/O Completion Port (IOCP), which will then, in turn, be used to create a background thread on the thread pool in order to serve the incoming APC. Which handle would it be exactly and when was this exactly done?

Thanks in advance!

https://blog.stephencleary.com/2013/11/there-is-no-thread.html#comment-d6183a70-4107-11ea-ac15-038facef7728