(Staticman) Stephen Cleary: Ah, that context is helpful! It has been many years since I've worked with asynchronous WCF services, but let's see...
`CallContext.Logical*Data` should work fine, especially if you need to use code across legacy and modern.

> I'm not clear on how *not* having an `async` method up the stack could cause this?

It's because an `async` method triggers the "copy on write" behavior for the logical call context.

> In my own code there is no `async` indeed

Ah, in that case the fix may be much simpler: ensure your code always calls `CallContext.FreeNamedDataSlot` to *delete* the data when it is done with it.

This step is redundant for `async` methods because when `CallContext.LogicalSetData` is called in an `async` method[1], the current logical call context is copied (if it hasn't been already), and when that `async` method completes, the previous logical call context is restored. So any adds/updates in the `async` method are removed/reverted automatically.

[1] - Technically, this refers to the closest `async` method going up the call stack. So if an `async` method calls a synchronous method that modifies the logical call context, the copied logical call context is considered the "current" one until the *parent* `async` method completes. Synchronous methods allow the logical call context values to flow back up the call stack, and since your code is synchronous, that may explain the behavior you're seeing.

> `MessageInspector` is registered in WCF pipeline as a singleton

I don't think that would be related. The "illogical" call contexts are tied to threads, and the logical call contexts are related to call stacks; neither of them have anything to do with object lifetimes.

> how does one "set the logical call context"

Sorry for the confusion, I meant just modifying the logical call context in any way, e.g., calling `CallContext.LogicalSetData`.

https://blog.stephencleary.com/2013/04/implicit-async-context-asynclocal.html#comment-fc93c4c0-41da-11ea-a7f0-f30dc1ba009d