(Staticman) Stephen Cleary: > for different incoming requests - it still should run in different `ExecutionContext`s. Is that assumption correct?
I don't know. That behavior is determined by the WCF runtime, and I don't know it to that level of detail.

It's possible that the runtime is always creating a new copy of the `ExecutionContext` for every request that comes in. It's also possible that the runtime is (at least for synchronous requests) using a shared `ExecutionContext` for efficiency.

https://blog.stephencleary.com/2013/04/implicit-async-context-asynclocal.html#comment-a54ac280-468b-11ea-8709-e79a8cf0fe62