(Staticman) Stephen Cleary: That depends on what you mean by "ok".
When keeping background work in memory, you should use something to notify ASP.NET that the work is ongoing, but that's the bare minimum. It's still possible to lose work if ASP.NET (or the machine running it) dies. If that's acceptable for the work you're doing, then yes, registration with ASP.NET is all you need.

But if you have a critical piece of work that, once submitted, *cannot* be lost, then you need a true distributed solution: at the very least a reliable queue. That will always be the case for critical work.

https://blog.stephencleary.com/2014/06/fire-and-forget-on-asp-net.html#comment-b5bcd860-482a-11ea-aa1f-e3aa3fdd1c82